test: #324 split the Postgres-only builder tests out of query_unit_test.go
Round-3 observation: this PR grew query_unit_test.go to 496 lines, four short of
the 500-line limit — and the file-size guard cannot catch it, because
listNonTestSources excludes _test.go.

Move the two tests covering query_postgres_build.go into a file named after it.
query_unit_test.go drops to 441, the new file is 65. Pure move: the relocated
block is byte-identical to what was removed, verified by diff rather than by
reading, and both tests still run.

Not widening the guard to test files here: performance_test.go (575) and
consistency_test.go (525) already exceed 500, so that guard would be red on
arrival for reasons this PR does not own. Filed separately.

Refs #324
